+++ exportplan/context.py
@@ -5,7 +5,7 @@ from exportplan import data
 
 class ExportPlanDashboardPageContextProvider(AbstractPageContextProvider):
 
-    template_name = 'exportplan/export_plan_dashboard_page.html'
+    template_name = 'exportplan/dashboard_page.html'
 
     @staticmethod
     def get_context_data(request, page):