+++ tests/unit/learn/test_context.py
@@ -1,5 +1,6 @@
 import pytest
 
+from tests.unit.core.factories import CuratedListPageFactory
 from tests.unit.learn import factories
 from core.context import get_context_provider
 
@@ -8,9 +9,9 @@ from core.context import get_context_provider
 def test_lesson_page_context_provider(rf, domestic_homepage, user):
     request = rf.get('/')
     request.user = user
-    factories.TopicPageFactory(slug='topic-one', parent=domestic_homepage)
-    factories.TopicPageFactory(slug='topic-two', parent=domestic_homepage)
-    topic = factories.TopicPageFactory(slug='topic-three', parent=domestic_homepage)
+    CuratedListPageFactory(slug='topic-one', parent=domestic_homepage)
+    CuratedListPageFactory(slug='topic-two', parent=domestic_homepage)
+    topic = CuratedListPageFactory(slug='topic-three', parent=domestic_homepage)
     page = factories.LessonPageFactory(slug='lesson-one', parent=topic)
 
     provider = get_context_provider(request=request, page=page)